pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "th3cl4w"
version = "0.1.0"
description = "Control system for Unitree D1 robotic arm"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "transforms3d>=0.4.1",
    "matplotlib>=3.7.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "opencv-python>=4.8.0",
    "requests>=2.28.0",
    "pyzmq>=25.0",
    "pyyaml>=6.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "google-generativeai>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "web*", "shared*", "services*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = false
ignore_missing_imports = true
disable_error_code = [
    "no-any-return",
    "arg-type",
    "assignment",
    "name-defined",
    "operator",
    "attr-defined",
    "union-attr",
    "call-overload",
    "index",
    "var-annotated",
    "return-value",
    "misc",
    "no-redef",
    "return",
    "list-item",
    "call-arg",
]

[tool.black]
line-length = 100
target-version = ["py310"]
